version 0.1.50

+++ lib/data/provider/conversion_provider.dart
@@ -1,6 +1,7 @@
 import 'package:aqua/data/provider/fiat_provider.dart';
 import 'package:aqua/features/settings/settings.dart';
 import 'package:aqua/features/shared/shared.dart';
+import 'package:decimal/decimal.dart';
 
 final _conversionProvider = Provider.family
     .autoDispose<ConversionProvider, (Asset, int)>(
@@ -18,14 +19,14 @@ class ConversionProvider {
 }
 
 final _conversionFiatProvider = Provider.family
-    .autoDispose<ConversionFiatProvider, (Asset, double)>(
+    .autoDispose<ConversionFiatProvider, (Asset, Decimal)>(
         (ref, arguments) => ConversionFiatProvider(ref, arguments));
 
 class ConversionFiatProvider {
   ConversionFiatProvider(this._ref, this._arguments);
 
   final AutoDisposeProviderRef _ref;
-  final (Asset, double) _arguments;
+  final (Asset, Decimal) _arguments;
 
   Stream<String> _conversion() => _ref
       .read(fiatProvider)
@@ -43,11 +44,11 @@ final conversionProvider =
 });
 
 final _conversionFiatStreamProvider = StreamProvider.family
-    .autoDispose<String, (Asset, double)>((ref, arguments) async* {
+    .autoDispose<String, (Asset, Decimal)>((ref, arguments) async* {
   yield* ref.watch(_conversionFiatProvider(arguments))._conversion();
 });
 
 final conversionFiatProvider =
-    Provider.family.autoDispose<String?, (Asset, double)>((ref, arguments) {
+    Provider.family.autoDispose<String?, (Asset, Decimal)>((ref, arguments) {
   return ref.watch(_conversionFiatStreamProvider(arguments)).asData?.value;
 });
